requirements.txt
Django
djangorestframework
python-dotenv
django-cors-headers
psycopg2
whitenoise